Separate implicit return value mixin
